+++ .env.example
@@ -32,9 +32,13 @@ AI_TIMEOUT_OLLAMA_MS=12000
 AI_TIMEOUT_OPENAI_MS=60000
 
 # API Server
+# Host donde escucha la API
 HOST=0.0.0.0
+# Puerto de escucha de la API
 PORT=8000
+# Nivel de logging de la aplicación (DEBUG, INFO, etc.)
 LOG_LEVEL=INFO
+# Si vale 1, SQLAlchemy mostrará cada consulta ejecutada
 DEBUG_SQL=0
 
 # Auth y seguridad
@@ -48,15 +52,26 @@ SESSION_EXPIRE_MINUTES=1440
 # Se ignora en producción; allí siempre es true
 COOKIE_SECURE=false
 COOKIE_DOMAIN=
+# Orígenes permitidos para CORS, separados por coma. En producción
+# deben especificarse dominios exactos; en desarrollo se completan
+# automáticamente las variantes localhost/127.0.0.1
 ALLOWED_ORIGINS=http://localhost:5173,http://127.0.0.1:5173
 # Usuario administrador inicial; si no existe, se crea al ejecutar migraciones
 ADMIN_USER=admin
 ADMIN_PASS=REEMPLAZAR_ADMIN_PASS
+# Tamaño máximo de archivos que acepta la API (MB)
 MAX_UPLOAD_MB=8
+# Si es true, requiere autenticación basada en sesiones
 AUTH_ENABLED=true
+# Resultados máximos por página al listar productos
 PRODUCTS_PAGE_MAX=100
+# Cantidad de entradas por página al consultar historial de precios
 PRICE_HISTORY_PAGE_SIZE=20
+# Crea productos canónicos automáticamente al importar
 AUTO_CREATE_CANONICAL=true
+# Umbral de similitud para sugerencias difusas
 FUZZY_SUGGESTION_THRESHOLD=0.87
+# Número de sugerencias devueltas en búsquedas aproximadas
 SUGGESTION_CANDIDATES=3
-VITE_MAX_UPLOAD_MB=15 # Tamaño máximo de subida en MB para el frontend
+# Tamaño máximo de subida en MB para el frontend
+VITE_MAX_UPLOAD_MB=15
