+++ frontend/.env.example
@@ -1,1 +1,2 @@
+# URL base de la API usada por el frontend en desarrollo
 VITE_API_URL=http://127.0.0.1:8000
